Removed login game option
Since AuthServiceProvider uses a reference of players as the User model, it doesn't have a connection set by default so it looks for the players table on the laravel DB which doesn't exist, so we must simply hardcode preservation as the default database connection for players.

This is tested as working fine and makes logical sense.