Added multi-world login via middleware
Laravel and authentication with Laravel is really not designed for authenticating against multiple databases. Inserting is different because it's less magic, you just generate some insert statements and execute it, but fetching is different because it does a lot of magic wiring things up behind the scenes.

Authenticating against different databases is a very atypical usage pattern for web applications, so it's not surprising it's not supported out of the box by Laravel.

This is probably fine, but it would need heavy testing before we can think about merging it and we may never merge it.